backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
openai==1.51.2
anthropic==0.34.2
stripe==11.1.0
